+++ bin/varnishd/mgt/mgt.h
@@ -227,7 +227,7 @@ void STV_Init(void);
 /* mgt_vcc.c */
 void mgt_DumpBuiltin(void);
 char *mgt_VccCompile(struct cli *, struct vclprog *, const char *vclname,
-    const char *vclsrc, const char * const *vclsrcfiles, int C_flag);
+    const char *vclsrc, const char * const *vclsrcfiles, const char *vcl_path, int C_flag);
 
 void mgt_vcl_init(void);
 void mgt_vcl_startup(struct cli *, const char *vclsrc, const char *origin,