+++ bin/varnishd/mgt/mgt_vcc.c
@@ -57,6 +57,7 @@ struct vcc_priv {
 #define VCC_PRIV_MAGIC	0x70080cb8
 	const char	*vclsrc;
 	const char * const *vclsrcfiles;
+	const char	*vcl_path;
 	struct vsb	*dir;
 	struct vsb	*csrcfile;
 	struct vsb	*libfile;
@@ -103,7 +104,7 @@ vcc_vext_iter_func(const char *filename, void *priv)
 static void v_noreturn_ v_matchproto_(vsub_func_f)
 run_vcc(void *priv)
 {
-	struct vsb *sb = NULL;
+	struct vsb *sb = NULL, *vsb = NULL;
 	struct vclprog *vpg;
 	struct vcc_priv *vp;
 	struct vcc *vcc;
@@ -118,8 +119,17 @@ run_vcc(void *priv)
 	vcc = VCC_New();
 	AN(vcc);
 	VCC_Builtin_VCL(vcc, builtin_vcl);
-	VCC_VCL_path(vcc, mgt_vcl_path);
 	VCC_VMOD_path(vcc, mgt_vmod_path);
+	if (vp->vcl_path != NULL) {
+		vsb = VSB_new_auto();
+		AN(vsb);
+		VSB_printf(vsb, "%s:%s", vp->vcl_path, mgt_vcl_path);
+		AZ(VSB_finish(vsb));
+		VCC_VCL_path(vcc, VSB_data(vsb));
+		VSB_destroy(&vsb);
+	} else {
+		VCC_VCL_path(vcc, mgt_vcl_path);
+	}
 
 #define VCC_FEATURE_BIT(U, l, d)			\
 	VCC_Opt_ ## l(vcc, MGT_VCC_FEATURE(VCC_FEATURE_ ## U));
@@ -347,7 +357,7 @@ mgt_vcc_fini_vp(struct vcc_priv *vp, int leave_lib)
 
 char *
 mgt_VccCompile(struct cli *cli, struct vclprog *vcl, const char *vclname,
-    const char *vclsrc, const char * const *vclsrcfiles, int C_flag)
+    const char *vclsrc, const char * const *vclsrcfiles, const char *vcl_path, int C_flag)
 {
 	struct vcc_priv vp[1];
 	struct vsb *sb;
@@ -362,6 +372,7 @@ mgt_VccCompile(struct cli *cli, struct vclprog *vcl, const char *vclname,
 	mgt_vcc_init_vp(vp);
 	vp->vclsrc = vclsrc;
 	vp->vclsrcfiles = vclsrcfiles;
+	vp->vcl_path = vcl_path;
 
 	/*
 	 * The subdirectory must have a unique name to 100% certain evade